requirements/base.txt
# Read .env files in Django.
python-dotenv ==0.9.1

# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.34.2

# Conservative Django
django>=1.11.29, <2.0

# Configuration
django-environ>=0.4.5, <0.5

# Forms
django-crispy-forms>=1.7.2, <1.8

# Models
django-model-utils>=3.2.0, <3.3

# Images
Pillow>=7.0.0, <7.1

# Password storage
argon2-cffi>=19.2.0, <19.3

# Python-PostgreSQL Database Adapter
psycopg2-binary>=2.8.4, <2.9

# Unicode slugification
awesome-slugify>=1.6.5, <1.7

# Time zones support
pytz>=2019.3

# Versioning
versioneer>=0.18

# Online documentation via django-docs.
Sphinx>=2.4.4, <2.5
django-docs>=0.3.1, <0.4
sphinx-rtd-theme>=0.4.3, <0.5

######################
# Project app imports
######################

# Django-plugins (with Django 1.11 support)
-e git+https://github.com/mikkonie/django-plugins.git@1bc07181e6ab68b0f9ed3a00382eb1f6519e1009#egg=django-plugins

# Rules for permissions
rules>=2.2, <2.3

# REST framework
djangorestframework>=3.11.0, <3.12

# Keyed list addon for DRF
drf-keyed-list==0.0.2

# Token authentication
django-rest-knox>=4.1.0, <4.2

# Markdown field support
markdown>=3.2.1, <3.3
django-markupfield>=1.5.1, <1.6
django-pagedown>=1.0.6, <1.1  # NOTE: v2.x requires Django 2.1+
mistune>=0.8.4, <0.9

# Background Jobs requirements
celery<4.2  # pyup: <4.2  # blocker: https://github.com/celery/celery/issues/4878

django-postgres-copy==2.3.5

-e git+https://github.com/bihealth/sodar_core.git@globalbgjobs#egg=django-sodar-core

# Simplejson is more advanced than then built-in one.
simplejson >=2.1

# Useful helper attrs.
attrs >=19.1.0
cattrs
python-dateutil

# Writing of Excel (XLSX) files.
xlsxwriter ==1.1.2

# Writing of VCF files
vcfpy
pysam==0.15.2

# Filter query generation
sqlalchemy ==1.2.13

# SQL Alchemy and Django integrate
SQLAlchemy ==1.2.13
aldjemy ==0.9.1

sqlparse

# Kenneth Reitz's wonderful HTTP library.
requests ==2.22.0
# Parsing of HTML
beautifulsoup4==4.8.1

# Numpy and friends
numpy ==1.14.3
pandas ==0.23.4

# Simple object proxies.
wrapt == 1.10.11

# Interval binning
interval-binning ==1.0.0

# Natural language processing toolkit
nltk

# Parsing of ISA-tab
altamisa ==0.2.5

# Login as every user
django-su

# Sentry client.
sentry-sdk ==0.14.0

raven ==6.10.0

# Redis support
redis==3.3.8

# HTTP Proxy for forwarding AJAX requests.
django-http-proxy==0.4.3

# YAML parsing
PyYAML

django-encrypted-model-fields==0.5.8
requests-http-signature==0.2.0
djangorestframework-httpsignature==1.0.0
httpsig==1.3.0
cryptography ==3.4.6
